Extract list of vars to replace from file
Instead of having a fixed set of variables that can
be replaced by replaceVars, extract the list of variables
from the file in which variables are to be replaced.
This makes replaceVars much more flexible, removing the
need to update it every time a new variable needs to be
replaced in some file.