Add modem firmware leaf package
Instead of passing the modem firmware file to the Makefile
on the command-line or in the environment, fetch the firmware's
leaf package and get the firmware image from that.  Also add
that firmware package dependency to the mangOH master leaf
package so that it's more readily available to developers
and so the mangOH .spk file build works without extra leaf
configuration steps.

Change WP77_RELEASE_VER to WP77_YOCTO_VER for clarity.